+++ src/hlib/libcpp/std.cc
@@ -1,5 +1,24 @@
 typedef std::string string;
 
+// should support all compilers(todo)
+// #if __GNUC__ || __clang__
+// typedef __int128 int128;
+// #else
+// typedef __int128_t int128;
+// #endif
+
+typedef long long int64;
+typedef unsigned long long uint64;
+
+typedef unsigned int uint;
+typedef unsigned long long uint64;
+
+typedef short int16;
+typedef unsigned short uint16;
+
+typedef unsigned char uint8;
+typedef char int8;
+
 void error(string err_msg){
 	std::cerr << err_msg << std::endl ;
 }